Renames, create mock_component.hpp
