+++ .clang-tidy
@@ -1,5 +1,5 @@
 Checks: ' -*, bugprone-*, cert-*, clang-analyzer-core.*, clang-analyzer-cplusplus.*, clang-analyzer-security.*, cppcoreguidelines-pro-bounds-pointer-arithmetic, cppcoreguidelines-pro-type-cstyle-cast, hicpp-* '
 WarningsAsErrors: 'bugprone-*,cert-*,clang-analyzer-*,hicpp-*'
 CheckOptions:
-- key: 'hicpp-special-member-functions.AllowSoleDefaultDtor'
-- value: 1
+  - key: 'hicpp-special-member-functions.AllowSoleDefaultDtor'
+  - value: 1
